refactor: extract validation logic from service
