+++ domain/build.gradle
@@ -4,6 +4,14 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.4'
 }
 
+bootJar {
+    enabled = false
+}
+
+jar {
+    enabled = true
+}
+
 repositories {
     mavenCentral()
 }
