+++ domain/src/main/java/com/qrlogi/domain/inspection/validator/InspectionValidator.java
@@ -1,0 +1,11 @@
+package com.qrlogi.domain.inspection.validator;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class InspectionValidator {
+
+
+}
