+++ domain/src/main/java/com/qrlogi/domain/orderitem/OrderItemValidator/OrderItemValidator.java
@@ -1,0 +1,21 @@
+package com.qrlogi.domain.orderitem.OrderItemValidator;
+
+import com.qrlogi.domain.orderitem.entity.OrderItem;
+import com.qrlogi.domain.orderitem.repository.OrderItemRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class OrderItemValidator {
+
+    private final OrderItemRepository orderItemRepository;
+
+    public OrderItem validateByOrderItemIdExists(Long orderItemId) {
+
+        return orderItemRepository.findById(orderItemId)
+                .orElseThrow(() -> new IllegalArgumentException("OrderItem not found"));
+
+
+    }
+}
