+++ domain/src/main/java/com/qrlogi/domain/product/entity/Product.java
@@ -3,10 +3,7 @@ package com.qrlogi.domain.product.entity;
 import com.qrlogi.domain.product.dto.ProductDTO;
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
@@ -34,7 +31,6 @@ public class Product {
     private String category;
 
 
-    //수정
     public void restoreProduct(ProductDTO productDTO) {
         if(productDTO.getName() != null) this.name = productDTO.getName();
         if(productDTO.getModelCode() != null) this.modelCode = productDTO.getModelCode();
