+++ domain/src/main/java/com/qrlogi/domain/product/service/ProductService.java
@@ -3,6 +3,7 @@ package com.qrlogi.domain.product.service;
 import com.qrlogi.domain.product.dto.ProductDTO;
 import com.qrlogi.domain.product.entity.Product;
 import com.qrlogi.domain.product.repository.ProductRepository;
+import com.qrlogi.domain.product.validator.ProductValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,6 +15,7 @@ import java.util.List;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final ProductValidator productValidator;
 
     @Transactional
     public ProductDTO registerProduct(ProductDTO productDTO) {
@@ -30,9 +32,7 @@ public class ProductService {
 
     public ProductDTO updateProduct(Long productId, ProductDTO productDTO) {
 
-        Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new IllegalArgumentException("없는 상품입니다."));
-
+        Product product = productValidator.validateProductExists(productId);
         product.restoreProduct(productDTO);
 
         return ProductDTO.toDTO(productRepository.save(product));
@@ -41,11 +41,13 @@ public class ProductService {
 
 
     public void deleteProduct(Long productId) {
+
         productRepository.deleteById(productId);
     }
 
 
     public List<ProductDTO> getAllProducts() {
+
         return productRepository.findAll().stream()
                 .map(ProductDTO::toDTO)
                 .toList();
@@ -53,7 +55,7 @@ public class ProductService {
 
 
     public ProductDTO getProductById(Long productId) {
-        return ProductDTO.toDTO(productRepository.findById(productId)
-                .orElseThrow(() -> new IllegalArgumentException("없는 상품입니다.")));
+
+        return ProductDTO.toDTO(productValidator.validateProductExists(productId));
     }
 }
