+++ domain/src/main/java/com/qrlogi/domain/shipment/service/ShippingService.java
@@ -4,6 +4,7 @@ import com.qrlogi.domain.orderitem.entity.OrderItem;
 import com.qrlogi.domain.orderitem.repository.OrderItemRepository;
 import com.qrlogi.domain.shipment.dto.ShippingRequest;
 import com.qrlogi.domain.shipment.dto.ShippingResponse;
+import com.qrlogi.domain.shipment.validator.ShipmentValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,11 +13,11 @@ import org.springframework.stereotype.Service;
 public class ShippingService {
 
     private final OrderItemRepository orderItemRepository;
+    private final ShipmentValidator shipmentValidator;
 
     public ShippingResponse doShipping(ShippingRequest shippingRequest) {
-        OrderItem item = orderItemRepository.findById(shippingRequest.getOrderItemId())
-                .orElseThrow(() -> new IllegalArgumentException("주문 항목이 존재하지 않습니다."));
 
+        OrderItem item =  shipmentValidator.validateShippmentExists(shippingRequest.getOrderItemId());
 
         return ShippingResponse.builder()
                 .orderItemId(item.getId())
@@ -26,7 +27,7 @@ public class ShippingService {
                 .containerSerial(shippingRequest.getContainerSerial())
                 .inspector1(shippingRequest.getInspector1())
                 .inspector2(shippingRequest.getInspector2())
-                .qc(item.getShippedQty() >= item.getOrderedQty()) //수량검수통과여부
+                .qc(item.getShippedQty() >= item.getOrderedQty()) //qc
                 .build();
 
     }
